Change `RecentArticleAction` format
Recent article action is now defines as a structure with one property type of `Observable<Void>` which represents when view will appear.

In this commit, there is also a subclass of `UIViewController` called `ViewLifeCycleViewController`: this class can be use as parent view controller to obtain a publish subject, binded on the view life cycle.